GDHI-88: Refactor Graph Components and style changes <Arjun/Shaswat>
